--- src/components/examples-page/index.js
@@ -1,3 +1,0 @@
-import main from './examples-page';
-
-export default main;
