+++ src/components/help-page/index.js
@@ -1,0 +1,3 @@
+import main from './help-page';
+
+export default main;
